+++ src/Symfony/Component/Notifier/Bridge/LineNotify/CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 ---
 
  * Add the `ssl` DSN option to send requests over plain HTTP
+ * Deprecate the bridge as LINE Notify was shut down, use the LineBot bridge instead
 
 6.3
 ---